requirements.txt
Django==5.2
mysqlclient==2.2.4
Pillow==11.3.0
PyYAML
